requirements.txt
torch
transformers
datasets
tokenizers
numpy
pandas
tqdm
python-dotenv
indic-nlp-library